jwt authentication

+++ controllers/api/v1/posts_api.js
@@ -24,11 +24,15 @@ module.exports.destroy =async function(req, res){
     try{
         const currPost = await Post.findById(req.params.id);
         // .id means converting the object id into string
-        currPost.deleteOne();
-        await Comment.deleteMany({
-            post: req.params.id,
-        })
-        return res.json(200, {message: "Post and associated comments deleted"});
+        if(post.user == req.user.id){
+            currPost.deleteOne();
+            await Comment.deleteMany({
+                post: req.params.id,
+            })
+            return res.json(200, {message: "Post and associated comments deleted"});
+        }else{
+            return res.json(401, {message: "You cannot delete this post"})
+        }
     }catch(err){
         console.log(err);
         return res.json(500, {message: "Internal server error"})
